refactor: refactor mail config and fix docker file
