+++ src/Life-Balance.WebApp/appsettings.json
@@ -8,7 +8,13 @@
   },
   "ConnectionStrings": {
     "SqlConnections": "Server=(localdb)\\MSSQLLocalDB;Initial Catalog=LifeBalanceDatabase;Integrated Security=True;Connect Timeout=30;Encrypt=False;TrustServerCertificate=False;ApplicationIntent=ReadWrite;MultiSubnetFailover=False",
-    "PostgreConnections": "Host=localhost;Port=5432;Database=LifeBalance;Username=postgres;Password=Aa23680!1"
+    "PostgreConnections": "connection string for postgre SQL database"
+  },
+  "MailSettings": {
+    "Server": "smtp.gmail.com",
+    "Port": "587",
+    "EmailAddress": "email@email.com",
+    "Password": "StrongPassword"
   },
   "AllowedHosts": "*"
 }
